==> 1_dataPreprocessing.ipynb <==
# Downloading Libraries


! pip install nltk
! pip install vaderSentiment
! pip install pytrends
! pip install textblob
! pip install wordcloud
! pip install gensim
! pip install seaborn
! pip install TextBlob
! pip install joblib

# Importing Libraries


import nltk
import logging
import numpy as np
import pandas as pd
import seaborn as sns
from textblob import TextBlob
import matplotlib.pyplot as plt
from nltk.corpus import stopwords

import re
import string
import joblib
import matplotlib.pyplot as plt

import warnings
warnings.filterwarnings('ignore')

# Download NLTK resources (if not already downloaded)
nltk.download('punkt')
nltk.download('stopwords')
nltk.download('wordnet')
nltk.download('wordnet', quiet=True)

# Importing Dataset


df = pd.read_csv('Dataset/Uncleaned Data/clean_reviews.csv')
df.head()

# Data Preprocessing


df.columns

df.columns

df.shape

df.info()

df.describe()

df.isnull().sum()

df.duplicated().sum()

# remove the following columns id, date and reviewer_id
df = df.drop(['id', 'reviewer_id', 'listing_id'], axis=1)
df.head()

def remove_urls(text):
    # Input validation
    if not isinstance(text, str):
        logging.error("Input must be a string.")
        raise TypeError("Input must be a string.")

    # Removing URLs from the text
    url_pattern = r'http\S+|www\S+|https\S+'
    cleaned_text = re.sub(url_pattern, '', text, flags=re.MULTILINE)

    return cleaned_text

def remove_mentions_hashtags(text):
    # Input validation
    if not isinstance(text, str):
        logging.error("Input must be a string.")
        raise TypeError("Input must be a string.")

    # Removing mentions and hashtags from the text
    pattern = r'\@\w+|\#\w+'
    cleaned_text = re.sub(pattern, '', text)

    return cleaned_text

def remove_punctuations(text):

    # Input validation
    if not isinstance(text, str):
        logging.error("Input must be a string.")
        raise TypeError("Input must be a string.")

    # Removing punctuations from the text
    translation_table = str.maketrans('', '', string.punctuation)
    cleaned_text = text.translate(translation_table)

    return cleaned_text

# Preload stop words to avoid reloading them every time the function is called
stop_words = set(stopwords.words('english'))


def remove_stopwords(text):
    # Input validation
    if not isinstance(text, str):
        logging.error("Input must be a string.")
        raise TypeError("Input must be a string.")

    # Removing stop words from the text
    filtered_words = [word for word in text.split() if word not in stop_words]
    cleaned_text = ' '.join(filtered_words)

    return cleaned_text

def remove_numbers(text):
    """Remove numbers from a text string"""
    return re.sub(r'\d+', '', text)

def remove_emojis(text):
    """Remove emojis from a text string"""
    emoji_pattern = re.compile(pattern="["
                               u"\U0001F600-\U0001F64F"  # emoticons
                               u"\U0001F300-\U0001F5FF"  # symbols & pictographs
                               u"\U0001F680-\U0001F6FF"  # transport & map symbols
                               u"\U0001F700-\U0001F77F"  # alchemical symbols
                               u"\U0001F780-\U0001F7FF"  # Geometric Shapes Extended
                               u"\U0001F800-\U0001F8FF"  # Supplemental Arrows-C
                               u"\U0001F900-\U0001F9FF"  # Supplemental Symbols and Pictographs
                               u"\U0001FA00-\U0001FA6F"  # Chess Symbols
                               u"\U0001FA70-\U0001FAFF"  # Symbols and Pictographs Extended-A
                               u"\U00002702-\U000027B0"  # Dingbats
                               u"\U000024C2-\U0001F251"
                               "]+", flags=re.UNICODE)
    return emoji_pattern.sub(r'', text)

def handle_whitespace(text):
    """Handle excessive whitespace"""
    return re.sub(r'\s+', ' ', text).strip()

def lemmitization(text):
    """Lemmitize text"""
    lemmatizer = nltk.stem.WordNetLemmatizer()
    return ' '.join([lemmatizer.lemmatize(word) for word in text.split()])

def stemminization(text):
    """Stem text"""
    stemmer = nltk.stem.PorterStemmer()
    return ' '.join([stemmer.stem(word) for word in text.split()])

def clean_text(text):
    text = text.lower()
    text = remove_urls(text)
    text = remove_mentions_hashtags(text)
    text = remove_punctuations(text)
    text = remove_stopwords(text)
    text = remove_numbers(text)
    text = remove_emojis(text)
    text = handle_whitespace(text)
    text = lemmitization(text)
    text = stemminization(text)

    # if new functions are added, add them here

    return text

# convert comments column to str type
df['comments'] = df['comments'].astype(str)

df['cleaned_comments'] = df['comments'].apply(clean_text)
df.head()

df.duplicated().sum()

# remove duplicates
df = df.drop_duplicates()
df.duplicated().sum()

# Sentiment analysis
df['polarity'] = df['cleaned_comments'].apply(
    lambda x: TextBlob(x).sentiment.polarity)
df['sentiment'] = pd.cut(df['polarity'], bins=3, labels=[
                         'negative', 'neutral', 'positive'])

# print the number of positive, negative and neutral reviews
df['sentiment'].value_counts()

# take 3915 positive reviews and 3915 negative reviews and 3915 neutral reviews
df = df.groupby('sentiment').apply(lambda x: x.sample(n=3915)).reset_index(
    drop=True).sort_values(by=['sentiment'])
df['sentiment'].value_counts()

df = pd.read_csv('Dataset/Cleaned Data/reviews.csv')
df.head()

# remove any rows with null values
df = df.dropna()
df.isnull().sum()

df.duplicated().sum()

# drop any rows which do not start with a number in the date column
df = df[df['date'].str.contains('^\d', na=False)]

df['date'] = pd.to_datetime(df['date'])

# store the cleaned data
df.to_csv('Dataset/Cleaned Data/reviews.csv', index=False)
