==> 4_LLM.ipynb <==
# Downloading Libraries

! pip install nltk
! pip install vaderSentiment
! pip install pytrends
! pip install textblob
! pip install wordcloud
! pip install gensim
! pip install seaborn
! pip install TextBlob
! pip install joblib
! pip install transformers

# Importing Libraries

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from transformers import BertTokenizer, TFBertForSequenceClassification
from tensorflow.keras.optimizers import Adam
import tensorflow as tf

# Importing Dataset

from google.colab import drive
drive.mount('/content/drive')

df = pd.read_csv('/content/drive/MyDrive/Datasets/Cleaned Data/reviews.csv')
df.head()

# Date Preprocessing

df.isnull().sum()

# Dropping the null values
df.dropna(inplace=True)
df.isnull().sum()

## **<center><u>Assessment: Experimentations (using Deep Learning)</u></center>**


# Initialize tokenizer
tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')

# Tokenize and encode sequences in the dataset
tokens = tokenizer.batch_encode_plus(
    df['cleaned_comments'].tolist(),
    max_length=128,  # You can adjust this depending on your text length
    padding='max_length',
    truncation=True,
    add_special_tokens=True,
    return_attention_mask=True,
    return_tensors='tf'
  )

# Prepare input data
input_ids = tokens['input_ids']
attention_mask = tokens['attention_mask']

# Prepare labels
df['sentiment'] = df['sentiment'].astype('category')
labels = df['sentiment'].cat.codes

input_ids_np = input_ids.numpy()
attention_mask_np = attention_mask.numpy()

labels_np = labels.to_numpy()

train_input_ids, test_input_ids, train_attention_mask, test_attention_mask, train_labels, test_labels = train_test_split(
    input_ids_np,
    attention_mask_np,
    labels_np,  # assuming labels_np is already a numpy array
    test_size=0.2,
    random_state=42
)

train_input_ids = tf.convert_to_tensor(train_input_ids)
train_attention_mask = tf.convert_to_tensor(train_attention_mask)
train_labels = tf.convert_to_tensor(train_labels)

test_input_ids = tf.convert_to_tensor(test_input_ids)
test_attention_mask = tf.convert_to_tensor(test_attention_mask)
test_labels = tf.convert_to_tensor(test_labels)

train_labels = tf.cast(train_labels, dtype=tf.int32)

model = TFBertForSequenceClassification.from_pretrained(
    'bert-base-uncased', num_labels=len(df['sentiment'].cat.categories))

optimizer = Adam(learning_rate=2e-5, epsilon=1e-08)
model.compile(optimizer='adam',
              loss='sparse_categorical_crossentropy', metrics=['accuracy'])

history = model.fit(
    {'input_ids': train_input_ids, 'attention_mask': train_attention_mask},
    train_labels,
    batch_size=32,  # Adjust based on your GPU memory
    epochs=10,  # BERT typically requires only a few epochs
    validation_split=0.1  # Use a part of the training set for validation
)

result = model.evaluate(
    x={'input_ids': test_input_ids, 'attention_mask': test_attention_mask},
    y=test_labels
)
print(f"Test loss: {result[0]}, Test accuracy: {result[1]}")
